fix infinite loop when counter compared with 1000

+++ drivers/gpu/drm/amd/display/dc/dce110/dce110_compressor.c
@@ -143,7 +143,7 @@ static void wait_for_fbc_state_changed(
 	struct dce110_compressor *cp110,
 	bool enabled)
 {
-	uint8_t counter = 0;
+	uint16_t counter = 0;
 	uint32_t addr = mmFBC_STATUS;
 	uint32_t value;
 
@@ -164,8 +164,6 @@ static void wait_for_fbc_state_changed(
 	} else {
 		DC_LOG_SYNC("FBC status changed to %d", enabled);
 	}
-
-
 }
 
 void dce110_compressor_power_up_fbc(struct compressor *compressor)
